properly organised js
